Parameter Make.3-IO

type t

val v : ?⁠flush_callback:(unit -> unit) -> fresh:bool -> generation:int64 -> fan_size:int64 -> string -> t

val v_readonly : string -> (t, [ `No_file_on_disk ]) Stdlib.result

val offset : t -> int64

val read : t -> off:int64 -> len:int -> bytes -> int

val clear : generation:int64 -> ?⁠hook:(unit -> unit) -> reopen:bool -> t -> unit

val flush : ?⁠no_callback:unit -> ?⁠with_fsync:bool -> t -> unit

val get_generation : t -> int64

val set_fanout : t -> string -> unit

val get_fanout : t -> string

val rename : src:t -> dst:t -> unit

val append : t -> string -> unit

val close : t -> unit

val size_header : t -> int

size_header t is t's header size.

module Lock : sig ... end

module Header : sig ... end

val exists : string -> bool

exists name is true iff there is a pre-existing IO instance called name.

val size : t -> int

Returns the true size of the underlying data representation in bytes. Note that this is not necessarily equal to the total size of observable data, which is given by offset.

To be used for debugging purposes only.