Module Private.Data

exception Invalid_size of string

The exception raised when trying to encode a key or a value of size other than encoded_size

module type Key = sig ... end

module type Value = sig ... end

module Entry : sig ... end

module String_fixed (L : sig ... end) : sig ... end