Module Data.String_fixed

Parameters

module L : sig ... end

Signature

type t = string

include Key with type t := string

val equal : string -> string -> bool

The equality function for keys.

val hash : string -> int

Note: Unevenly distributed hash functions may result in performance drops.

val hash_size : int

The number of bits necessary to encode the maximum output value of hash. `Hashtbl.hash` uses 30 bits.

Overestimating the hash_size will result in performance drops; underestimation will result in undefined behavior.

include Value with type t := string

val t : string Repr.t

val encode : string -> string

val encoded_size : int

val decode : string -> int -> string