Module Private.Io_array

module type ELT = sig ... end

module type S = sig ... end

module Make = Index__.Io_array.Make

Takes an IO instance and wraps it in an Array interface with support for prefetching sections of the array.