Module type Index_intf.Index

module Key : sig ... end

module Value : sig ... end

module type IO = sig ... end

module type SEMAPHORE = sig ... end

module type THREAD = sig ... end

module Cache : sig ... end

Signatures and implementations of caches. Make requires a cache in order to provide instance sharing.

module type S = sig ... end

Index module signature.

exception RO_not_allowed

The exception raised when a write operation is attempted on a read_only index.

exception RW_not_allowed

The exception is raised when a sync operation is attempted on a read-write index.

exception Closed

The exception raised when any operation is attempted on a closed index, except for close, which is idempotent.

module Make : functor (K : Key.S) -> functor (V : Value.S) -> functor (IO : IO) -> functor (_ : SEMAPHORE) -> functor (T : THREAD) -> functor (C : Cache.S) -> S with type key = K.t and type value = V.t

module Stats : sig ... end

Run-time metric tracking for index instances.

module Checks = Index.Checks

module Private : sig ... end

These modules should not be used. They are exposed purely for testing purposes.