Module Io.Extend

Parameters

S : S

Signature

include S with type t = S.t

type t = S.t

val v : ?⁠flush_callback:(unit -> unit) -> fresh:bool -> generation:Index__.Import.int63 -> fan_size:Index__.Import.int63 -> string -> t

val v_readonly : string -> (t, [ `No_file_on_disk ]) Stdlib.result

val offset : t -> Index__.Import.int63

val read : t -> off:Index__.Import.int63 -> len:int -> bytes -> int

val clear : generation:Index__.Import.int63 -> ?⁠hook:(unit -> unit) -> reopen:bool -> t -> unit

val flush : ?⁠no_callback:unit -> ?⁠with_fsync:bool -> t -> unit

val get_generation : t -> Index__.Import.int63

val set_fanout : t -> string -> unit

val get_fanout : t -> string

val get_fanout_size : t -> Index__.Import.int63

val rename : src:t -> dst:t -> unit

val append : t -> string -> unit

val append_substring : t -> string -> off:int -> len:int -> unit

val close : t -> unit

val size_header : t -> int

size_header t is t's header size.

module Lock : sig ... end

module Header : sig ... end

val exists : string -> bool

exists name is true iff there is a pre-existing IO instance called name.

val size : t -> int

Returns the true size of the underlying data representation in bytes. Note that this is not necessarily equal to the total size of observable data, which is given by offset.

To be used for debugging purposes only.

val iter : page_size:Index__.Import.int63 -> ?⁠min:Index__.Import.int63 -> ?⁠max:Index__.Import.int63 -> (off:Index__.Import.int63 -> buf:string -> buf_off:int -> int) -> t -> unit