Module Index__Io_array.Make

Takes an IO instance and wraps it in an Array interface with support for prefetching sections of the array.

Parameters

IO : Index__.Io.S

Elt : ELT

Signature

include Index__.Search.ARRAY

type t

type elt

val get : t -> Index__.Import.int63 -> elt

val length : t -> Index__.Import.int63

val pre_fetch : t -> low:Index__.Import.int63 -> high:Index__.Import.int63 -> unit

type io = IO.t

val v : io -> t