Module Make.Checks

Offline fsck-like utility for checking the integrity of Index stores built using this module.

val cli : unit -> Index__.Checks_intf.empty

Run a Cmdliner binary containing tools for running offline integrity checks.