Module Index_unix__Buffer

Extensible buffers with non-allocating access to the buffer's contents.

type t

The type of buffers.

val create : int -> t

create n is a fresh buffer with initial size n.

val length : t -> int

length b is the number of bytes contained in the buffer.

val is_empty : t -> bool

is_empty t iff t contains 0 characters.

val clear : t -> unit

clear t clears the data contained in t. It does not reset the buffer to its initial size.

val add_substring : t -> string -> off:int -> len:int -> unit

add_substring t s ~off ~len appends the substring s.(off) .. s.(off + len - 1) at the end of t, resizing t if necessary.

val add_string : t -> string -> unit

add_string t s appends s at the end of t, resizing t if necessary.

val write_with : (string -> int -> int -> unit) -> t -> unit

write_with writer t uses writer to write the contents of t. writer takes a string to write, an offset and a length.