Module Index_unix__Raw

Raw wraps a file-descriptor with an file-format used internally by Index. The format contains the following header fields:

offset: a 64-bit integer, denoting the length of the file containing valid data;

version: an 8-byte version string;

generation: a 64-bit integer denoting the generation number;

fan: a 64-bit length field, followed by a string containing that many bytes.