Module Private.Io_array

module type ELT = sig ... end

module type S = sig ... end

module Make (IO : sig ... end) (Elt : ELT) : S with type io = IO.t and type elt = Elt.t

Takes an IO instance and wraps it in an Array interface with support for prefetching sections of the array.