Module Io_array.Make

Takes an IO instance and wraps it in an Array interface with support for prefetching sections of the array.

Parameters

module IO : sig ... end

module Elt : ELT

Signature

type t

type elt = Elt.t

val get : t -> Optint.Int63.t -> elt

val length : t -> Optint.Int63.t

val pre_fetch : t -> low:Optint.Int63.t -> high:Optint.Int63.t -> unit

type io = IO.t

val v : io -> t